Human friendly progress reporting
Telepresence now uses a progress reporter that is very similar to the
one used by Docker compose. The implementation is a variation of that
reporter's source code, so big thanks to the Docker Compose CLI
authors for making it available as OSS.

A new global `--progress <progress>` flag was added. It defaults to
"auto" which means that the style chosen depends on if the command runs
from a tty type terminal or not. Other possible values are "plain",
"quiet", and "json". `--progress quiet` is implied when formatted output
 is chosen using `--output json|yaml`.

Signed-off-by: Thomas Hallgren <thomas@tada.se>